switch initial level selection configuration to .toml file
