+++ .gitignore
@@ -2,8 +2,6 @@
 
 Cargo.lock
 
-src/level/setup.rs
-
 **/.DS_Store
 
 **/ldtk.AppImage
