+++ Cargo.toml
@@ -8,6 +8,8 @@ bevy = { version = "0.15.0", features = ["dynamic_linking"] }
 bevy_ecs_ldtk = "0.11.0"
 bevy_rapier2d = "0.28.0"
 enum-map = "2.7.3"
+serde = { version = "1.0.217", features = ["derive"] }
+toml = "0.8.19"
 
 [profile.dev]
 opt-level = 1
