+++ Lightborne.toml
@@ -1,0 +1,3 @@
+[level_config]
+level_index = 3
+level_path = "levels/lightborne.ldtk"
